upscaled fast tower soldiers
In the future perhaps the fast tower should have a machine gun integrated into the tower, leading to faster fire, where as the basic tower would just be a nest for the soldiers. Also it could be cool if the houses weren't just sitting there, and we could station soldiers in police stations for example, or snipers on top of houses.